Fix bugs, add example
