+++ README.md
@@ -1,7 +1,7 @@
 # flask-djangoify
 To us Flask in Django way
 ## Introduction
-The orignal idea form [simple-payment](https://github.com/Aalto-LeTech/simple-payment) by [Jaakko Kantojärvi](https://github.com/raphendyr)
+The orignal idea form [simple-payments](https://github.com/Aalto-LeTech/simple-payments) by [Jaakko Kantojärvi](https://github.com/raphendyr)
 ## TODO
 - Include authencation, based on flask-login
 - Make a flask extension
