+++ sample/test/__init__.py
@@ -1,0 +1,1 @@
+from .view import bp
