+++ sample/test/view.py
@@ -1,0 +1,8 @@
+from flask_djangoify.utils import Blueprint
+
+bp= Blueprint('test', __name__, url_prefix='/test')
+
+
+@bp.route('/', methods=['GET', 'POST'])
+def test_view():
+    return 'hello test'
